pyproject.toml
[project]
name = "chatbotlitellm"
version = "0.1.0"
description = "Intelligent chatbot using Chainlit and LiteLLM with Google Palm-2 via OpenRouter"
readme = "README.md"
authors = [
    { name = "Muhammad Hammad", email = "hafizhammad1883@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.13.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "google-auth>=2.23.0",
    "google-cloud-aiplatform>=1.36.0",
]

[project.scripts]
chatbotlitellm = "chatbotlitellm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
